requirements.txt
google-genai
python-dotenv
streamlit
gradio-client
huggingface_hub
matplotlib
seaborn